Point is now a struct, and its components (X and Y) now have public get accessors (I don't know why they didn't before).
Path now has two new properties, Length and PathVector, both of which have only get accessors and do not function as variables. Length returns the length of the path as a float, plain and simple. PathVector returns the vector that defines the path (so, the differences of the components of the start and end points), and it returns this as a Point.

GamePiece has a new property, MapPos. Like Length and PathVector, this has only a get accessor and does not function as a variable. MapPos returns, in the form of a Point, the position of the piece in map-space coordinates.