Added operator overrides to the Point class, we can now add and subtract Points as well as multiply points by floats. Everything works the same as it does in math graph. I've adjusted the Path and GamePiece classes to reflect this change.
Finished the Intersects() method in the Path class.

CalculateIntersections() and Parse() in the Map class are both functional, and the Map class is now complete enough to allow for instantiation and movement. I made a simple map file (it's in bin\Windows GL\debug) and added some code to the Game1 class to load it in - feel free to remove this code, it was only used for testing purposes.

I think that's all.