Added the static method DrawMap to the Map class. This method takes a Map and a GraphicsDevice and returns a Texture2D that *should* serve as a background image for the map. Presently it only draws lines where the paths should be, but we can alter it to draw more complicated shapes if we want, and we can still use hand-drawn backgrounds if we want to.
Because I needed a GraphicsDevice to work with Texture2D objects, Menu's constructor now takes a GraphicsDevice. Also, there is now a Texture2D variable in Menu for storing the map background. I'm not familiar with any of the game's drawing code, so someone else will need to make the necessary changes for working this texture into the game.