The Map class will now check map files for pellets as well as paths. A pellet will be denoted by a single coordinate pair in the map file, and can be inserted at any location in the file.
The Path class now contains a list of all the GamePieces on the path, and all GamePieces now have a PacmanCollision method. I've added a section to the Update method in Menu to check all the GamePieces on pacman's current path for collisions with pacman. If one is detected, that object's PacmanCollision method will be called. As for implementation of these PacmanCollision methods, I've done the following:
Ghost calls Pacman.GhostCollision, a new method in the Pacman class. GhostCollision has not been implemented.
Pellet's PacmanCollision method has not been implemented.

I've expanded on Ghost's GetNextDirection method. It will now ensure that the chosen next direction actually allows for a valid path change, and it allows the ghost to turn around if they, by some miracle, reach the end of a path and stop.