Added big pellet functionality. This entailed creating a BigPellet class, creating a static BigPellet method in Menu, and making the Menu object in Game1 static so that the BigPellet method could be referenced from within BigPellet's collision method.
Big pellet parsing is also in the map class now - we can make a pellet into a big pellet by adding "b" to the end of the ordered pair in the map file. There's one big pellet in the map file as of right now, use that as an example.