Fixed the big pellet issue with light levels.
Pacman will now spawn at the first path in the map, ghosts at the second. I adjusted the map file accordingly.

Map scaling is officially complete. Not only will the textures be drawn correctly to the map size, but the map size itself will be adjusted such that all mobile game objects will move at the same speed on every map, and such that every map will have the same dimensions. Offsets are now universal, so we can drop in any map file we want and it'll run perfectly. I also added a bunch of error trapping to the Parse method - if an error is found on any line in the map file, the line will simply be discarded. There's no longer any chance of crashing the program with a bad map file.

The addition of true map scaling now causes all map-based points to be created with float values rather than int values, and this initially resulted in paths not properly connecting and pellets not being properly created. The == operator override in the point class has been altered to allow for a very small tolerance - points that are within .05 tiles of each other are considered to be equal. A similar alteration was added to Path.PointOnPath, for the same reason - any point within .05 tiles of a path is considered to be on the path. Given that all maps will now be 50 tiles by 50 tiles thanks to scaling, I think this tolerance is small enough that it will never cause a problem, and if necessary I think we could reduce it even further without issue.